Add JNI backend target which does not allow valgrind
There are a lot of tests which override valgrind requests or disable themselves
if valgrind is enabled. This makes it so that the test can be confugured to use
JNI backend but not allow valgrind without having to skip if valgrind is
enabled or call an override method.

In LocalCluster deprecate overrideAnyRequestForValgrind() in favor of using
BackendTarget.NATIVE_EE_JNI_NO_VG